xds: Add a KeyValueStore-backed xDS delegate extension to contrib
An xDS delegate extension point was added in
https://github.com/envoyproxy/envoy/pull/22473 to enable custom behavior
upon receiving and loading xDS resources.  This change creates an
implementation of the XdsResourcesDelegate interface that is backed by a
KeyValueStore.

The intended use case is to enable persisting xDS resources and loading
them on startup in Envoy Mobile, in the event that the xDS control plane
is unreachable.

Signed-off-by: Ali Beyad <abeyad@google.com>